+++ sdk/python/clean-pypi.sh
@@ -1,1 +1,1 @@
-pypi-cleanup -u flet -p flet -y -r '0.6.0.dev12' --do-it
+pypi-cleanup -u flet -p flet-core -y -r '0.11.0.dev' --do-it